SpikingRNNs

Documentation for SpikingRNNs.

SpikingRNNs.forward_signal!

SpikingRNNs.lognorm_reparametrize

SpikingRNNs.wmat_train_assemblies_protocol

SpikingRNNs.forward_signal! — Method

send_signal(conn::ConnectionStateRate)

Computes the input to postsynaptic population, given the current state of presynaptic population. For a rate model, it applies the iofunction to the neuron potentials, gets the rate values then multiplies rates by weights, adding the result to the input of the postsynaptic population.

SpikingRNNs.lognorm_reparametrize — Method

    lognorm_reparametrize(m,std) -> distr::LogNormal

parameters

m sample mean

std sample std

SpikingRNNs.wmat_train_assemblies_protocol — Method

wmat_train_assemblies_protocol(Nneus::Integer,p::Real; scal::Float64=1.0) 
  -> (weights::SparseMatrixCSC,assembly_dictionary::Dict)

Arguments

Npost::Integer : number of postsynaptic neurons

Nassemblies::Integer : number of assemblies to train

p_assembly::Real : probability for a neuron to be in any assembly

scal::Real=1.0 : connection weights

Outputs

weights : sparse weight matrix

assembly_list::Vector{NamedTuple{(:neupost,:neupre),Vector{Int64,Vector{Int64}}}} element i contains the pre and post neurons associated to assembly i