SpikingRNNs

Warning

The documentation is still missing. Please see the "examples" section for usage.

Examples

Leaky-integrate and fire, single neurons – Voltage traces, refractoriness, etc.

SpikingRNNs.PSIFNeuron

SpikingRNNs.forward_signal!

SpikingRNNs.hardbounds

SpikingRNNs.lognorm_reparametrize

SpikingRNNs.next_poisson_spiketime

SpikingRNNs.next_poisson_spiketime_from_function

SpikingRNNs.PSIFNeuron — Method

PSIFNeuron(n::Integer,τ::R,cap::R,
    v_threshold::R,v_reset::R,v_leak::R,t_refractory::R) where R

Generates a population of leaky integrate and fire neurons.

Arguments

n : number of neurons

τ : dynamics time constant

cap : capacitance

v_threshold : treshold for action potential emission

v_reset : reset value after firing

v_leak : leak potential

t_refractory : duration of refractory period

SpikingRNNs.forward_signal! — Method

send_signal(conn::ConnectionStateRate)

Computes the input to postsynaptic population, given the current state of presynaptic population. For a rate model, it applies the iofunction to the neuron potentials, gets the rate values then multiplies rates by weights, adding the result to the input of the postsynaptic population.

SpikingRNNs.hardbounds — Method

hardbounds(x::R,low::R,high::R) where R = min(high,max(x,low))

Applies hard-bounds on scalar x

SpikingRNNs.lognorm_reparametrize — Method

lognorm_reparametrize(m::Real,std::Real) -> d::LogNormal

Parameters

m::Real sample mean

std::Real sample std

Returns

d::Distributions.LogNormal

SpikingRNNs.next_poisson_spiketime — Method

nextpoissonspiketime(tcurrent::Float64,rate::Float64) -> tnext::Float64

Returns next spike after current time t_current in a random Poisson process. with rate rate.

SpikingRNNs.next_poisson_spiketime_from_function — Method

nextpoissonspiketimefromfunction(tcurrent::Float64,funrate::Function,funrateupper::Function; Tmax::Float64=0.0,nowarning::Bool=false) -> Float64

Returns the next spiketime in a Poisson process with time-varying rate. The rate variation is given by function fun_rate.

See e.g. Laub,Taimre,Pollet 2015

Arguments

t_start::Float64 : current time

fun_rate::Function : fun_rate(t::Float64) -> r::Float64 returns rate at time t

fun_rate_upper::Function : upper limit to the function above. Strictly decreasing in t must be as close as possible to the fun_rate for efficiency

Tmax::Float64 : upper threshold for spike proposal, maximum interval that can be produced

nowarning::Bool : does not throw a warning when Tmax` is reached