SpikingRNNs

Warning

The documentation is still missing. Please see the "examples" section for usage.

Examples

Leaky-integrate and fire, single neurons – Voltage traces, refractoriness, etc.

Leaky-integrate and fire, sinusoidal input – Voltage traces, refractoriness, etc.

Leaky-integrate and fire, voltage density – Current based IF, voltage distribution in a single neuron, for either stationary input, or noisy input

Triplet rule for two poisson neurons

Index and documented functions

SpikingRNNs.PSIFNeuron

SpikingRNNs.PairSTDP

SpikingRNNs.PlasticityTriplets

SpikingRNNs.RecWeightsFull

SpikingRNNs.bin_spikes

SpikingRNNs.draw_spike_raster

SpikingRNNs.forward_signal!

SpikingRNNs.hardbounds

SpikingRNNs.lognorm_reparametrize

SpikingRNNs.next_poisson_spiketime

SpikingRNNs.next_poisson_spiketime_from_function

SpikingRNNs.sparse_constant_wmat

SpikingRNNs.PSIFNeuron — Method

PSIFNeuron(n::Integer,τ::R,cap::R,
    v_threshold::R,v_reset::R,v_leak::R,t_refractory::R) where R

Generates a population of leaky integrate and fire neurons.

Arguments

n : number of neurons

τ : dynamics time constant

cap : capacitance

v_threshold : treshold for action potential emission

v_reset : reset value after firing

v_leak : leak potential

t_refractory : duration of refractory period

SpikingRNNs.PairSTDP — Type

function PairSTDP(τplus,τminus,Aplus,Aminus,n_post,n_pre;
     plasticity_bounds=PlasticityBoundsNonnegative())

constructor for "classic" pairwise STPD rule. The Aminus coefficient, when positive, causes an increase in the weights. So it should be set negative for classic Hebbian STDP

SpikingRNNs.PlasticityTriplets — Type

PlasticityTriplets(τplus,τminus,τx,τy, A2plus,A3plus,A2minus,A3minus,npost,npre; plasticity_bounds=PlasticityBoundsNonnegative())

it construc!

SpikingRNNs.RecWeightsFull — Type

RecWeightsFull(conn::Connection, everydt::Float64,Tend::Float64;Tstart::Float64=0.0)

Constructor, fields are self explanatory

SpikingRNNs.bin_spikes — Method

bin_spikes(Y::Vector{R},dt::R,Tend::R;Tstart::R=0.0) where R

Arguments

Y::Vector{<:Real} : vector of spike times

dt::Real : time bin size

Tend::Real : end time for the raster

Optional argument

Tstart::Real=0.0 : start time for the raster

Returns

binned_spikes::Vector{<:Integer} : binned_spikes[k] is the number of spikes that occur in the timebin k (i.e. between Tstart + (k-1)*dt and Tstart + k*dt)

SpikingRNNs.draw_spike_raster — Method

drawspikeraster(trains::Vector{Vector{Float64}}, dt::Real,Tend::Real; Tstart::Real=0.0, spikesize::Integer = 5, spikeseparator::Integer = 1, backgroundcolor::Color=RGB(1.,1.,1.), spikecolors::Union{C,Vector{C}}=RGB(0.,0.0,0.0), max_size::Real=1E4) where C<:Color

Draws a matrix that contains the raster plot of the spike train.

Arguments

Trains : Vector of spike trains. The order of the vector corresponds to the order of the plot. First element is at the top, second is second row, etc.

dt : time interval representing one horizontal pixel

Tend : final time to be considered

Optional arguments

Tstart::Real : starting time

spike_size::Integer : heigh of spike (in pixels)

spike_separator::Integer : space between spikes, and vertical padding

background_color::Color : self-explanatory

spike_colors::Union{Color,Vector{Color}} : if a single color, color of all spikes, if vector of colors, color for each neuron (length should be same as number of neurons)

max_size::Integer : throws an error if image is larger than this number (in pixels)

Returns

raster_matrix::Matrix{Color} you can save it as a png file

SpikingRNNs.forward_signal! — Method

send_signal(conn::ConnectionStateRate)

Computes the input to postsynaptic population, given the current state of presynaptic population. For a rate model, it applies the iofunction to the neuron potentials, gets the rate values then multiplies rates by weights, adding the result to the input of the postsynaptic population.

SpikingRNNs.hardbounds — Method

hardbounds(x::R,low::R,high::R) where R = min(high,max(x,low))

Applies hard-bounds on scalar x

SpikingRNNs.lognorm_reparametrize — Method

lognorm_reparametrize(m::Real,std::Real) -> d::LogNormal

Parameters

m::Real sample mean

std::Real sample std

Returns

d::Distributions.LogNormal

SpikingRNNs.next_poisson_spiketime — Method

nextpoissonspiketime(tcurrent::Float64,rate::Float64) -> tnext::Float64

Returns next spike after current time t_current in a random Poisson process. with rate rate.

SpikingRNNs.next_poisson_spiketime_from_function — Method

nextpoissonspiketimefromfunction(tcurrent::Float64,funrate::Function,funrateupper::Function; Tmax::Float64=0.0,nowarning::Bool=false) -> Float64

Returns the next spiketime in a Poisson process with time-varying rate. The rate variation is given by function fun_rate.

See e.g. Laub,Taimre,Pollet 2015

Arguments

t_start::Float64 : current time

fun_rate::Function : fun_rate(t::Float64) -> r::Float64 returns rate at time t

fun_rate_upper::Function : upper limit to the function above. Strictly decreasing in t must be as close as possible to the fun_rate for efficiency

Tmax::Float64 : upper threshold for spike proposal, maximum interval that can be produced

nowarning::Bool : does not throw a warning when Tmax` is reached

SpikingRNNs.sparse_constant_wmat — Method

sparseconstantwmat(npost::Integer,npre::Integer,p::Real,jval::Real ; scal::Float64=1.0, noautapses::Bool=true, rowsum::Union{Nothing,Float64}=nothing)

Sparse weight matrix

Arguments

npost::Integer : number of postsynaptic neurons

npre::Integer : number of presynaptic neurons

p::Real : matrix density

j_val::Real : constant weight in matrix

Optional arguments

scal::Float64=1.0 : at the very end scales the matrix by this value

no_autapses::Bool : removes autapses

rowsum::Union{Nothing,Float64}=nothing : if a number, each row sums to this number (before scaling)