SpikingRNNs

Documentation for SpikingRNNs.

SpikingRNNs.lognorm_reparametrize

SpikingRNNs.send_signal

SpikingRNNs.lognorm_reparametrize — Method

    lognorm_reparametrize(m,std) -> distr::LogNormal

parameters

m sample mean

std sample std

SpikingRNNs.send_signal — Method

send_signal(conn::ConnStateRate)

Computes the input to postsynaptic population, given the current state of presynaptic population. For a rate model, it applies the iofunction to the neuron potentials, gets the rate values then multiplies rates by weights, adding the result to the input of the postsynaptic population.